Add the minimum required elements to the gemspec
See comments in the gemspec file. spec.files, spec.extensions, and the
spec.require_paths are the gotchas to watch out for (IME).

Also, changed the license to be the same as the c library.

Also, remove the `Rb` module name that `bundler gem` decided I needed,
because I put a `-rb` on the end of the gem name. 🙄