Add VldMail::Validation class
In the Init_* functions for the module and the class, call
rb_define_module and rb_define_class_under.

It compiles now, and you can instantiate a class that does not
much. Added the test, which will change later and found I needed a
relative require for the bundle. Also, I added the bundle (an artifact
of the compile) to the .gitignore, so it doesn't get checked in.

At this stage, when you need to experiment with the C implementation,
you can `cd ext/vldmail`, run `ruby extconf.rb`, and run `make`. In
the same directory, you can fire up `irb` and `require_relative
'./vldmail`, and you're able to play with the C implementation in a
REPL. Well, assuming it compiles :-)