Add VldMailValidation#initialize
In Init_vldmail_validation (the ruby initializer for the class),
define an initialize method on the ruby object.

Using ruby's Check_Type, the associated c function asserts that its
input is a string, raising an error otherwise.

We convert the input string to a multibyte string (char *). Then,
because libvldmail expects it, we convert the multibyte string to a
wide character string.

We make a new pointer to a libvldmail struct. We call Data_Get_Struct
to bind the struct we allocated during instantiation to our new
pointer. Then we call libvldmail's validate_email function, passing it
the wide character email string. The results are copied to the
location pointed to by our validation pointer.

The instance (self) now wraps the struct returned by libvldmail's
validate_email function, though we cannot as of yet get at its
contents from ruby.

We return self.