Add Validation#message
This method returns the libvldmail message. On success, the message
will be an empty string. On failure, it will be a validation message.

Use rb_define_method in Init_vldmail_validation to make the ruby
method.

In the c message function, get the libvldmail c struct, wrapped by the
class. Convert the libvldmail wide character message string to a
multi-byte string, using the wcstombs function.

For some reason, the libvldmail message either ends in a newline or has
one added somehow by the call to wctombs. To "strip" it, we use
strcspn to get the index of the first (only) newline. We then set that
element of the string to a null terminator. Et voila.

Finally, use rb_str_new_cstr to make a ruby VALUE representation of a
string. Return that.