Handle input that exceeds maximum length
One of the email RFCs gives 320 as the max length from an
email. libvldmail uses that as the length for its own email input. We
have to do something to prevent overflowing the buffer.

We test the length of the ruby input, using RSTRING_LEN. If it is
within limit, validate the email as usual. If it is over, set the
libvldmail struct values to our own error.