Add Validation#email
To make the Validation class more developer-friendly, add an email
attribute to store the input email string.

In Init_vldmail_validation, create an attr_reader on the ruby object.

In Validation#initialize, use rb_intern to obtain an ID (the c
equivalent of the ruby Symbol object) for the name of an ivar,
i.e. `:"@email"` in ruby-land. Then use the rb_ivar_set function to
set the ivar to the ruby email string on the instance of Validation.